+++ TSDF_forGS/cameras.h
@@ -1,11 +1,11 @@
 #include <glm/glm.hpp>
 
-glm::vec2 projectPointToPixel(const double x, const double y, const double z, const glm::mat4x3& P);
+glm::vec2 projectPointToPixel(const float x, const float y, const float z, const glm::mat4x3& P);
 //project a 3D point to pixel
 
-double get_value(const double* image, double u, double v, int width, int height);
+float get_value(const float* image, float u, float v, int width, int height);
 //get the corresponding value of (u,v) on the image
 
-void Linear_Interp(double& x, double& y, double& z, const double& x1, const double& y1, const double& z1, const double& x2, const double& y2, const double& z2, const double& value1, const double& value2);
+void Linear_Interp(float& x, float& y, float& z, const float& x1, const float& y1, const float& z1, const float& x2, const float& y2, const float& z2, const float& value1, const float& value2);
 
-double get_dist(const double x1, const double y1, const double z1, const double x2, const double y2, const double z2);
+float get_dist(const float x1, const float y1, const float z1, const float x2, const float y2, const float z2);
